src/requirements.txt
fastapi
jinja2
sqlalchemy
pyjwt
alembic
psycopg2-binary
python-multipart
bcrypt
